travis: Cleaner logs
This patch moves some of the steps to prepare the tests into the
pre-scripts.
This has the effect of cleaning up the log display a little bit and
will focus more on the test results.
Even if the logs are hidden, travis would still fail if any of the
steps fails.

Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>