[WIP] Start adding cloning tests
This is a WIP PR to help add some cloning tests.  I've created some
utility functions to help create a PVC and a Pod which will run a bash
script while attached to a PVC (for checking the PVC).  Right now we're
not actually testing cloning but just getting some of the machinations
nailed down.

Test is not passing yet due to bad shell when checking PVC.

Requires a cluster with a default storage class, ie.
   export KUBEVIRT_PROVIDER=os-3.10.0

Signed-off-by: Adam Litke <alitke@redhat.com>