Fix clone tests (#386)
* test: fix pvc annotation check

The clone controller uses a slightly different annotation on PVCs.  The
test was using the form used by the import controller.  Quick fix.

Signed-off-by: Adam Litke <alitke@redhat.com>

* tests: Pods must be polled to determine their node

When checking the node on which a pod is scheduled we must use an
Eventually assertion and poll.  Otherwise, tests could flake when pod
scheduling is delayed.

Signed-off-by: Adam Litke <alitke@redhat.com>