Fix lint errors in controller pkg, add lint check
Rather than dump a giant change set in one shot, we're going to tackle
bringing CDI up to lint standards bottoms up package by package.

This change just focuses on the controller package and is mostly just
adding proper comment strings on exported functions and constants.
We'll work on fixing up constant names in common and fix some of the
unnecessarily exported constants when we get to that step.