Fix cluster-clean
We were missing some resources in the cleanup on make cluster-clean
(when running against an external cluster).  We also weren't patching
any of the finalizers so the result is a delete of the cdi namespace
would hang in terminating state forever.

This change explicitly checks all resources on the cluster for a cdi
label and it also patches the cdi objects to clear out the finalizer.

Closes #774