Disable caching when converting qcow2 images
When converting images the qemu-img command uses a writeback cache mode
by default.  This means that writes to a block device go through the
host page cache and are lazily flushed to the storage device.  When
using this cache mode, the process can exit before all writes have
reached storage and our DataVolume could appear ready to use before all
I/O has completed.  This becomes a problem with shared storage because a
second host does not have the benefit of the first host's page cache
when reading.  To prevent this problem we must perform I/O directly to
the storage device.  This behavior is selected by passing "-t" "none" to
qemu-img.

This problem will not arise with "create", "info", or "resize" qemu-img
commands because those are metadata operations.

The following document provides some good background on qemu cache modes:
https://documentation.suse.com/sles/11-SP4/html/SLES-kvm4zseries/cha-qemu-cachemodes.html

Signed-off-by: Adam Litke <alitke@redhat.com>