Use long timeout for uploadproxy client (#1139)
The timeout value used by the uploadproxy client sets an upper limit on
the amount of time an upload can take.  After this limit is reached the
proxy will close the connection with the uploadserver resulting in a
failed upload.  We do want a reasonable timeout to prevent zombie upload
connections from persisting forever but we don't want to prevent
legitimate operations.  Increase the timeout value to 24 hours which
should be enough even for reasonably large images over relatively slow
connections.  Uploads that take longer than one day should probably be
done with an import process anyway.

Signed-off-by: Adam Litke <alitke@redhat.com>