Add critical addons toleration to cdi deployment
By toleratin the CriticalAddonsOnly taint, we ensure that our components
will get a reserved spot for critical addons in case they are being
evicted. This will reduce of chance of CDI component/s starving for a
place on a node and permanently unavailable.

Also added a functional test to verify this behaviour.

Signed-off-by: Daniel Belenky <dbelenky@redhat.com>