ux: Clarify message when PVC is too small (#1162)
When a user imports into a PVC that is not large enough we print an
error stating that "shrink not supported".  When this situation happens
it is far more likely that the user simply used a PVC that is too small
to hold the image.  Let's make it easier for them to debug by suggesting
this.  The verbage about shrinking is simply confusing.

Signed-off-by: Adam Litke <alitke@redhat.com>