Eliminate the need for HEAD requests in the Go HTTP download client. (#1122)
* Tolerate failing to find a Content-Length via HEAD requests.

Some servers don't accept HEAD requests, and Content-Length is
an optional field anyway.
In such cases, qemu-img won't be able to do a streaming
conversion of the URL, so download using our Go download code
to scratch space and convert the resulting file.
Use the field "brokenForQemuImg" to note this scenario.

This fixes creating a DataVolume using one of our examples,
manifests/example/import-kubevirt-datavolume.yaml

Signed-off-by: Maya Rashish <mrashish@redhat.com>

* Don't bother with basic auth, this test isn't about that

Signed-off-by: Maya Rashish <mrashish@redhat.com>

* Test for reasonable behaviour if HEAD requests are forbidden

Make sure that we mark it broken for qemu-img, but still succeed
in continuing.

Signed-off-by: Maya Rashish <mrashish@redhat.com>

* Use %v to print err

Signed-off-by: Maya Rashish <mrashish@redhat.com>

* Fix linter failure

Signed-off-by: Maya Rashish <mrashish@redhat.com>