Fix expectation in test that kills container, executor might be killed -9 by parent.
It might happen that `kill 1` is executed so fast that the parent process tries to exit and kills
the child process - `bash` which actually runs the `kill 1` command. In effect our process exits
with code 137. In case of this test it is not a problem, we expect process to kill parent.

Signed-off-by: Bartosz Rybacki <brybacki@redhat.com>