Make CDI infra deployments as critical addons. (#1361)
* Make CDI infra deployments as critical addons.

While here, mark cdi-operator as linux-only, it doesn't take
its value from the CDI CR.

Signed-off-by: Maya Rashish <mrashish@redhat.com>

* Add test CDI pods are deployed with critical addon tolerations

Co-authored-by: Daniel Belenky <dbelenky@redhat.com>
Signed-off-by: Daniel Belenky <dbelenky@redhat.com>
Signed-off-by: Maya Rashish <mrashish@redhat.com>

* Turn nfs-server into a deployment

So that it is restarted after the destructive test destroying
CDI testing pods.

Signed-off-by: Maya Rashish <mrashish@redhat.com>

* Rewrite CDI critical taint test

Now it is simpler and waits for pods to be running before
continuing for the next tests.

Signed-off-by: Maya Rashish <mrashish@redhat.com>

* Skip tolerations test if the CDI CR has unexpected values

This is necessary for Prow, which installs the CDI CR from the
previous version and updates to the latest code, but might be
useful in other cases too.

Signed-off-by: Maya Rashish <mrashish@redhat.com>

* Modify functional test to use AfterEach instead of defer

While here, make it more generic so there's a chance this Describe
block will be used by other tests - iterate over all pods instead
of just the CDI test pods.

Signed-off-by: Maya Rashish <mrashish@redhat.com>

* Skip test if CDI CR isn't named "cdi" (rather than fail)

Signed-off-by: Maya Rashish <mrashish@redhat.com>

* fix typo

Signed-off-by: Maya Rashish <mrashish@redhat.com>

Co-authored-by: Daniel Belenky <dbelenky@redhat.com>