Simplify the HTTP file host
Eliminate the part of the init container where we copy the images
to /tmp/shared/images, instead have the http container built with
the images into it.

This eliminates several places necessary to modify when adding
an image, an especially painful process until you remember that
the file-host container uses imagePullPolicy: IfNotPresent.

This step also created the tinyCore.qcow2 and tinyCore.qcow.gz files
using the RAW file, so we need to add it to the repository. A step
like this is useful to save space, but raw is the fattest format,
so we might be better off reversing it if we want to re-introduce it.

Signed-off-by: Maya Rashish <mrashish@redhat.com>