Wait for stray pods to terminate, destroy/re-create at AfterEach. (#1459)
This suite creates stray pods by deleting/re-creating the CDI CR,
causing several pods to terminate & be re-created.

This disrupts the critical addons test, which counts the number of
pods before and after.

To avoid doing this, mimic the critical addons test and wait until
the number of pods is the same as before the test run.

While here, also destroy/re-create the CDI CR unconditionally in
AfterEach. We want all the CDI pods to be running at the end of
a test, and being sure that they were first terminated lets us
be certain that we are checking that the pods are the new ones,
not old pods that haven't started terminating yet.

Signed-off-by: Maya Rashish <mrashish@redhat.com>