In the operator test there is a critical addons test that removes and (#1513)
restarts the pod used in tests. It verified the pods where 'ready' but
not that the service using those pods had working endpoints. Due to the
changes where we start the next test without waiting for the namespace
to finish deleting, those services didn't have enough time to have their
endpoints be available. This ensures the endpoints are available before
moving onto the next tests.

Signed-off-by: Alexander Wels <awels@redhat.com>