Use nbdkit for direct stream for the http importer  (#1508)
* Add nbdkit with qemu-img

Nbdkit can be used to stream directly the content of the image together
with qemu-img. It supports different plugins and filters.

Signed-off-by: Alice Frosi <afrosi@redhat.com>

* Use nbdkit to the http importer

Add nbdkit to stream directly the image content avoiding the scratch
space for the http importer.

Signed-off-by: Alice Frosi <afrosi@redhat.com>

* Adjust functional test cases with nbdkit

With the introduction of nbdkit, certain error messages do not match anymore,
and needed to be adjusted to the new behavior.

The scratch space is not needed anymore when the source is http for the majority of the cases.
Certain test cases checked the use of the scratch space, and those are not valid anymore.

Signed-off-by: Alice Frosi <afrosi@redhat.com>