Make clone test that checks fsGroup less flaky.
We start 3 verifier pods here:
1. Check the disk group id,	delete,	wait for termination
2. Check the target MD5,	delete,	don't wait for termination
3. Check the target MD5,	delete,	don't wait for termination
We then compare target MD5 with target MD5.

if #2 hasn't terminated yet, #3 might not start in time.

Let's swap #2 for a pod that checks the source MD5 - it makes more
sense to do that, and it also doesn't require us to wait for pods
to terminate.

Signed-off-by: Maya Rashish <mrashish@redhat.com>