Update kubevirtci (#1579)
* Update kubevirtci

Signed-off-by: Maya Rashish <mrashish@redhat.com>

* Sync kubevirtci gocli automatically, and move update logic to own script.

It's a separate script because we seem to only be able to pull in the
latest gocli tag, and we want update-codegen.sh to be reproducible.

The expected way to update is now:
Adjust git checksum in the file and run: ./hack/update-kubevirtci.sh

Signed-off-by: Maya Rashish <mrashish@redhat.com>

* Sync kubevirtci with new script.

Now that we don't run kubevirtci bump in bazel, we can see some files
were deleted long ago.

Also bumps KUBEVIRTCI_TAG.

Signed-off-by: Maya Rashish <mrashish@redhat.com>