Nbdkit refactor to unify usage for vddk and http importers (#1680)
* refactor: unify nbdkit usage

Nbdkit is used in the http and vddk importer and the code is completely
separated. This PR unifies the usages of nbdkit in a single file. For
the http importer, nbdkit was used together with qemu-img. With this
change nbdkit is started a separate process, and qemu-img read from the
local nbdkit socket.

Signed-off-by: Alice Frosi <afrosi@redhat.com>

* test: adjust the tests

Removed nbdkit together with qemu-img. Hence, we need to adjust the
tests to reflect the new behavior.

Signed-off-by: Alice Frosi <afrosi@redhat.com>