Remove test for invalid-qcow-backing-file.img
This is a malformed image - it gets treated as RAW image for
most steps, but as a corrupt VMDK file for the resize step.
Resize recognizes that the VMDK file is corrupt, and errors.

If resize doesn't happen, say we are importing to a block,
import is successful.

The presence of this test implies that the test should always
fail, but it doesn't, and is confusing.

Signed-off-by: Maya Rashish <mrashish@redhat.com>