Remove the test - delete scratch space (#1840)
In the current environment this test is unstable. The is a chaos
test where evil user or script tries to delete scratch space in
the moment where it is created but the import pod has not started yet.

The test brings more problems than real value so it is being removed.

Signed-off-by: Bartosz Rybacki <brybacki@redhat.com>