Add "AwaitingVDDK" back to condition reason. (#1816)
* Move AwaitingVDDK constant to common.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* Copy pending PVC bound condition reason to DV.

Replace the fixed "Pending" string and tweak the unit test that checked
this.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* Add a functional test for AwaitingVDDK.

Delete v2v-vmware ConfigMap and create a DataVolume, and the bound
condition should have a reason of "AwaitingVDDK".

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* Move AwaitingVDDK to its own functional test.

Retain common test code with existing table, but tack on an extra
cleanup step so v2v-vmware ConfigMap can be restored afterward.

Signed-off-by: Matthew Arnold <marnold@redhat.com>