Fix fieldRef apiVersion spam
Since #1864 we're seeing spam in the operator logs:
`[{\"op\":\"remove\",\"path\":\"/spec/template/spec/containers/0/env/6/valueFrom/fieldRef/apiVersion\"},`

This happens because we don't specify the fieldRef's apiVersion explicitly when creating the controller,
so k8s populates it for us -> causing us to reconcile it back.

Will now set the apiVersion explicitly to avoid this spam.

Signed-off-by: Alex Kalenyuk <akalenyu@redhat.com>