Fix flake causing test and move to destructive lane
This test deletes the CDIConfig and waits for it to recreate with a different UID.
However, this condition alone will not suffice as the CDIConfig status might still be empty for a few seconds,
causing havoc in following tests.

Will now wait for .status to not be empty, and marking test as destructive.

Signed-off-by: Alex Kalenyuk <akalenyu@redhat.com>