Fix 2 common flakes (#1898)
* Fix flake causing test and move to destructive lane

This test deletes the CDIConfig and waits for it to recreate with a different UID.
However, this condition alone will not suffice as the CDIConfig status might still be empty for a few seconds,
causing havoc in following tests.

Will now wait for .status to not be empty, and marking test as destructive.

Signed-off-by: Alex Kalenyuk <akalenyu@redhat.com>

* Fix second flake

Test waits for *any* error after making the HTTP request for upload from the host, which doesn't cut it;
That could be an EOF (witnessed), port-forward, etc -> while the only error that would guarantee the upload process
actually got triggered is `Unexpected return value 500`.
Without seeing this error we can't continue testing the scenario (can't be sure any upload logic even triggered).

Signed-off-by: Alex Kalenyuk <akalenyu@redhat.com>