Implement multi-stage ImageIO imports. (#1903)
* Add qemu-img rebase and commit operations.

Also only fail images with backing files that do not exist, so that
ImageIO snapshots can be downloaded and applied to a base disk image.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* Add merge phase to data processor.

This keeps qemu-img details out of the ImageIO data source.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* Beef up transfer ticket finalization/cancellation.

Snapshots seem to be more prone to getting locked indefinitely than
disks if not correctly finalized or cancelled, so do this more carefully
than before.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* Allow downloading snapshots from ImageIO.

Download the first snapshot as a raw whole-disk image, and download
subsequent snapshots as QCOW images to be committed to that base.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* Allow multi-stage fields on ImageIO data sources.

Also avoid removing base disk image when cleaning data directory.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* Add ImageIO multi-stage functional tests.

Pick up fakeovirt update for stub functionality, so inventory responses
can be changed on the fly for individual tests.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* Update multi-stage documentation for ImageIO.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* Move if-else test block to functions.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* Reset ImageIO inventory for a test I missed.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* Clean up from some review comments.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* Sort out calls to cleanupTransfer.

Failures during the creation of a transfer ticket call the original
cleanupTransfer in a single location, and any exits after the data
source is created call a wrapper function. The wrapper has a lock and a
'done' flag to make sure it is only called once on exit, even when
interrupted from the goroutine that waits for SIGTERM.

Signed-off-by: Matthew Arnold <marnold@redhat.com>