Explicitly set that storage class name
When the storage request is specified using the 'storage' API in
a DataVolume, CDI uses Storage Profiles to look up any missing
values that would be required to create the underlying PVC.
If the user omits the storage class parameter in their DV,
CDI assumes the cluster's default storage class will be used.

Since CDI calculates the storage class that should be used in
this case, it should explicitly set that storage class name
in the PVC it creates.

Signed-off-by: Bartosz Rybacki <brybacki@redhat.com>