Fix blank block preallocation import (#2005)
When trying to preallocate big size such as 1Gi
the preallocation failed since the dd occupied too
much memory and OOMKiller killed the process.
To fix it the dd will write in 1Mi block size the
requested size (will write the extra reminder if the
size is not aligned to 1Mi)
There was a test which checked blank block preallocation
but the size was only 100Mi which didnt fail, changed to 1Gi.

Signed-off-by: Shelly Kagan <skagan@redhat.com>