[release-v1.38] Backport fix hpp tests (#2047)
* Fix hpp setup for tests. (#2046)

With hpp operator at v0.11.0 it now requires cert-manager, and webhook
This PR modifies the hpp setup to include setting up those items.

Signed-off-by: Alexander Wels <awels@redhat.com>

* Update default storage to use CSI hpp driver (#1978)

* Update default storage to use CSI hpp driver

Update the default storage for hpp lanes to be the csi driver
instead of the legacy driver.

Signed-off-by: Alexander Wels <awels@redhat.com>

* Add check for csi driver capabilities in particular fsGroup.
Skip check for fsGroup if csi driver reports it cannot accomidate.

Signed-off-by: Alexander Wels <awels@redhat.com>

* Addressed review comment.

Signed-off-by: Alexander Wels <awels@redhat.com>