Imporve directory detection on containerdisk import (#2050)
Check the tar header metadata if a tar entry is a directory instead of
checking if it has a `/` suffix.

This solves the following issue:

While most applications write a `/` at the end of a directory entry in a
tar header, it is not necessarily there. If it is missing, the directory
will be treated as the target file and CDI tries and fails to import the
`disk` directory itself:

```
iI1203 12:02:26.740362       1 transport.go:149] File 'disk' found in the layer
I1203 12:02:26.740461       1 util.go:172] Writing data...
E1203 12:02:26.742269       1 registry-datasource.go:123] Error reading directory
E1203 12:02:26.742319       1 data-processor.go:229] readdirent: not a directory
image file does not exist in image directory
```

Signed-off-by: Roman Mohr <rmohr@redhat.com>