Use GetRange instead of Read in progress monitor.
Functional tests revealed a case where a QCOW image could be transferred
over the extents API if the server refuses to serve raw data. This is
likely a test problem only, but it's worth handling it in case it ever
really happens. The problem is that each Read from the progress monitor
will advance the reader's offset, when the actual image transfer is also
trying to Read, so it starts requesting incorrect ranges.

Signed-off-by: Matthew Arnold <marnold@redhat.com>