Allow http in s3 datasource
We are using the AWS s3 client, and if the URL we pass in starts
with http, the AddScheme function in the AWS S3 client replaces
the scheme with https. This checks the URL scheme and if it is
http, it disables TLS, which causes the AWS S3 client to set the
scheme internally as http.

Signed-off-by: Alexander Wels <awels@redhat.com>