Use ImageIO extents API to copy raw images more efficiently. (#2052)
* Add ImageIO extent reader to copy ranges of bytes.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* Use ImageIO extents API only if it is enabled.

Not all versions support the extents API, so check for the feature
before trying to create an extents reader.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* Copy extents to destination file.

Also break apart StreamDataToFile a bit so that StreamExtents can open
the destination the same way.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* Punch holes in block devices.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* Allow progress calculation over multiple readers.

This avoids setting the progress to 100% before all the extent readers
are really done. Each extent gets one reader, and without this change
the progress updater can see a random extent's Done flag go to true and
consider the whole transfer complete.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* Include zero extents in progress calculation.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* Minor file improvements for large disks.

Record any error code from Close, and try to do an fsync before the one
in importer.go. For some reason the first fsync always fails on larger
disks, so add this one to make sure the other one succeeds.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* Monitor extent progress to avoid download timeout.

Some configurations are unable to quickly punch holes or otherwise zero
out ranges in the destination file, so large disk transfers can trigger
timeouts on the oVirt side waiting for I/O. Monitor progress in the
background and extend the transfer ticket's validity with Extend and a
very small disk read. This keeps oVirt from canceling the download when
the importer is busy writing huge blocks of zero bytes to the
destination storage.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* Unit tests for checking extents feature existence.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* Minor changes from further unit test development.

* Allow different polling times for monitorExtentsProgress
* Make sure to close response body after HTTP requests
* Do not fail import if extents feature can't be checked

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* Move support code for extents unit tests.

Move these chunks so that all the Describe sections are together.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* More unit tests for ImageIO extents.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* Add unit tests for progress reporter changes.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* Make GetRange and Read work like other reads.

These needed to be more careful about reading the right amount of data,
and not returning unexpected EOFs. Functional tests work much better
now.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* Use GetRange instead of Read in progress monitor.

Functional tests revealed a case where a QCOW image could be transferred
over the extents API if the server refuses to serve raw data. This is
likely a test problem only, but it's worth handling it in case it ever
really happens. The problem is that each Read from the progress monitor
will advance the reader's offset, when the actual image transfer is also
trying to Read, so it starts requesting incorrect ranges.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* Functional tests for ImageIO extents.

The existing default test configuration actually already supports
extents, so add a test that explicitly disables them.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* Correctly zero ranges on block devices.

Handle configurations that cannot punch holes with fallocate by falling
back to regular writes, and make sure data copy is resumed at the
correct offset.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* Remove extra length check in GetRange.

This is not useful, it just gets the length of the error message
returned by the server and then returns the error message as the data to
write out.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* Use truncate instead of seek to append zeroes.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* Add unit tests for zero block utility functions.

Signed-off-by: Matthew Arnold <marnold@redhat.com>

* Add unit tests for some ImageIO failure cases.

ImageIO extents transfer should fail if the extents are not returned in
order, and if the server kills the session during the transfer.

Signed-off-by: Matthew Arnold <marnold@redhat.com>